CLI-623 Rename agent-integrate-install to agent-integrate-postlude
Per review: postlude pairs with the existing agent-integrate-prelude (opening
vs closing step), making the file's role obvious at a glance. Pure rename of
the file holding finalizeAgentInstall plus the cursor/codex/copilot imports.
