+++ src/cli/commands/integrate/cursor/index.ts
@@ -22,7 +22,7 @@
 
 import type { ResolvedAuth } from '../../../../lib/auth-resolver';
 import { warn } from '../../../../ui';
-import { finalizeAgentInstall } from '../_common/agent-integrate-install';
+import { finalizeAgentInstall } from '../_common/agent-integrate-postlude';
 import { displayAgentIntegratePrelude } from '../_common/agent-integrate-prelude';
 import { resolveSqaaSetup } from '../_common/sqaa-entitlement';
 import type { IntegrateAgentOptions } from '../_common/types';
